// src/users/dto/create-admin-users.dto.ts
export class CreateAdminUsersDto{
    name: string;

    user_type: string;

    staff_code: string;

    phone_number: string;
  
    email: string;

    password: string;
  
}

export class CreatesStudentDto{
    name: string;

    user_type: string;

    student_id: string;

    phone_number: string;
  
    email: string;

    password: string;
}